그룹 생성 로직 작성

+++ src/main/java/com/second/damoa/group/repository/GroupInfoRepository.java
@@ -3,5 +3,5 @@ package com.second.damoa.group.repository;
 import com.second.damoa.group.model.GroupInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupInfoRepository extends JpaRepository<GroupInfo, Integer> {
+public interface GroupInfoRepository extends JpaRepository<GroupInfo, Long> {
 }
